+++ src/DTO/Fields/InlineField/ImageField.php
@@ -4,9 +4,7 @@ namespace Ehyiah\QuillJsBundle\DTO\Fields\InlineField;
 
 use Ehyiah\QuillJsBundle\DTO\Fields\Interfaces\QuillFieldModuleInterface;
 use Ehyiah\QuillJsBundle\DTO\Fields\Interfaces\QuillInlineFieldInterface;
-use Ehyiah\QuillJsBundle\DTO\Modules\ImageAttributesModule;
-use Ehyiah\QuillJsBundle\DTO\Modules\ImageDragAndDropModule;
-use Ehyiah\QuillJsBundle\DTO\Modules\ResizeModule;
+use Ehyiah\QuillJsBundle\DTO\Modules\ImageSelectionModule;
 
 class ImageField implements QuillInlineFieldInterface, QuillFieldModuleInterface
 {
@@ -17,6 +15,6 @@ class ImageField implements QuillInlineFieldInterface, QuillFieldModuleInterface
 
     public static function importModules(): array
     {
-        return [ResizeModule::class, ImageDragAndDropModule::class, ImageAttributesModule::class];
+        return [ImageSelectionModule::class];
     }
 }