+++ tests/Form/QuillTypeTest.php
@@ -10,7 +10,8 @@ use Ehyiah\QuillJsBundle\DTO\Fields\InlineField\ImageField;
 use Ehyiah\QuillJsBundle\DTO\Fields\InlineField\ItalicField;
 use Ehyiah\QuillJsBundle\DTO\Fields\InlineField\UnderlineField;
 use Ehyiah\QuillJsBundle\DTO\Modules\EmojiModule;
-use Ehyiah\QuillJsBundle\DTO\Modules\ImageAttributesModule;
+use Ehyiah\QuillJsBundle\DTO\Modules\ImageSelectionModule;
+use Ehyiah\QuillJsBundle\DTO\Modules\NodeMoverModule;
 use Ehyiah\QuillJsBundle\DTO\Modules\ImageDragAndDropModule;
 use Ehyiah\QuillJsBundle\DTO\Modules\ResizeModule;
 use Ehyiah\QuillJsBundle\DTO\Modules\SyntaxModule;
@@ -94,10 +95,9 @@ final class QuillTypeTest extends TestCase
                 ],
                 'modules' => [
                     new SyntaxModule(),
-                    new ResizeModule(),
-                    new ImageDragAndDropModule(),
-                    new ImageAttributesModule(),
+                    new ImageSelectionModule(),
                     new EmojiModule(),
+                    new NodeMoverModule(),
                 ],
             ],
         ];
@@ -119,6 +119,7 @@ final class QuillTypeTest extends TestCase
                 'quill_extra_options' => [
                 ],
                 'modules' => [
+                    new NodeMoverModule(),
                 ],
             ],
         ];